fix build

+++ include/luisa/core/basic_traits.h
@@ -49,8 +49,9 @@ static_assert(std::is_arithmetic_v<half>,
 using uchar = uint8_t;
 using ushort = uint16_t;
 using uint = uint32_t;
-using ulong = uint64_t;
-using slong = int64_t;// long has different size on different platforms
+
+using ulong = unsigned long long;
+using slong = long long;// long has different size on different platforms
 
 using canonical_c_long = std::conditional_t<sizeof(long) == sizeof(int), int, slong>;
 using canonical_c_ulong = std::conditional_t<sizeof(long) == sizeof(int), uint, ulong>;
